LayoutTests/css3/filters/backdrop/backdropfilter-property-expected.txt

Test the parsing of the -webkit-backdrop-filter property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-backdrop-filter') is 'hue-rotate(120deg)'
PASS successfullyParsed is true

TEST COMPLETE
